+++ packages/api-client/src/granules.ts
@@ -87,6 +87,39 @@ export const getGranuleResponse = async (params: {
 };
 
 /**
+ * GET granule record with granuleGroup information from /granules/getGranuleAndGroup
+ *
+ * @param params - params
+ * @param params.prefix - the prefix configured for the stack
+ * @param params.body - param sent to endpoint, (granuleId of desired granule)
+ * @param params.callback - async function to invoke the api lambda
+ *                          that takes a prefix / user payload,
+ *                          cumulusApiClient.invokeApifunction
+ *                          is the default to invoke the api lambda
+ * @returns - the granule fetched by the API
+ */
+export const getGranuleAndGroup = async (params: {
+  prefix: string,
+  body: Object,
+  callback?: InvokeApiFunction
+}): Promise<ApiGatewayLambdaHttpProxyResponse> => {
+  const { prefix, body, callback = invokeApi } = params;
+  return await callback({
+    prefix: prefix,
+    payload: {
+      httpMethod: 'GET',
+      resource: '/{proxy+}',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      path: '/granules/getGranuleAndGroup',
+      body: JSON.stringify(body),
+    },
+    expectedStatusCodes: 202,
+  });
+};
+
+/**
  * GET granule record from /granules/{granuleId} or /granules/{collectionId}/{granuleId}
  *
  * @param params - params