+++ packages/db/src/index.ts
@@ -150,6 +150,7 @@ export {
   getApiGranuleCumulusIds,
   getApiGranuleExecutionCumulusIds,
   getGranuleCollectionId,
+  getGranuleAndGroupFromPG,
   getUniqueGranuleByGranuleId,
   getGranuleByUniqueColumns,
   upsertGranuleWithExecutionJoinRecord,