+++ packages/db/src/search/queries.ts
@@ -17,6 +17,7 @@ const reservedWords = [
   'infix',
   'estimateTableRowCount',
   'fields',
+  'includeActiveState',
   'includeFullRecord',
   'searchContext',
   'countOnly',
@@ -252,6 +253,7 @@ export const convertQueryStringToDbQueryParameters = (
     infix,
     fields,
     estimateTableRowCount,
+    includeActiveState,
     includeFullRecord,
     countOnly,
   } = queryStringParameters;
@@ -267,6 +269,7 @@ export const convertQueryStringToDbQueryParameters = (
   if (typeof fields === 'string') dbQueryParameters.fields = fields.split(',');
   dbQueryParameters.estimateTableRowCount = (estimateTableRowCount === 'true');
   dbQueryParameters.includeFullRecord = (includeFullRecord === 'true');
+  dbQueryParameters.includeActiveState = (includeActiveState === 'true');
   dbQueryParameters.countOnly = (countOnly === 'true');
   dbQueryParameters.sort = convertSort(type, queryStringParameters);