+++ packages/db/tests/search/test-GranuleSearch.js
@@ -13,12 +13,14 @@ const {
   fakeProviderRecordFactory,
   generateLocalTestDb,
   GranulePgModel,
+  GranuleGroupsPgModel,
   GranuleSearch,
   PdrPgModel,
   ProviderPgModel,
   migrationDir,
   FilePgModel,
   fakeFileRecordFactory,
+  fakeGranuleGroupRecordFactory,
   ExecutionPgModel,
   fakeExecutionRecordFactory,
   GranulesExecutionsPgModel,
@@ -219,6 +221,16 @@ test.before(async (t) => {
     t.context.pgGranules.map((granule, i) => ({
       granule_cumulus_id: granule.cumulus_id,
       execution_cumulus_id: executionRecords[99 - i].cumulus_id,
+    }))
+  );
+
+  t.context.granuleGroupsPgModel = new GranuleGroupsPgModel();
+  t.context.granuleGroups = await t.context.granuleGroupsPgModel.insert(
+    knex,
+    range(100).map((num) => fakeGranuleGroupRecordFactory({
+      granule_cumulus_id: num + 1,
+      group_id: num % 2 === 0 ? 2 : 1,
+      state: num % 5 === 0 ? 'A' : 'H',
     }))
   );
 });
@@ -1057,5 +1069,52 @@ test('GranuleSearch with archived: false pulls only non-archive granules', async
   const response = await dbSearch.query(knex);
   response.results.forEach((granuleRecord) => {
     t.is(granuleRecord.archived, false);
+  });
+});
+
+test('GranuleSearch with includeActiveStatus set to true returns GranuleGroup info', async (t) => {
+  const { knex } = t.context;
+  const queryStringParameters = {
+    includeActiveState: 'true',
+  };
+  const dbSearch = new GranuleSearch({ queryStringParameters });
+  const response = await dbSearch.query(knex);
+  response.results.forEach((granuleRecord) => {
+    t.true('state' in granuleRecord && (granuleRecord.state === 'A' || granuleRecord.state === 'H'));
+    t.true('groupId' in granuleRecord && (granuleRecord.groupId === 2 || granuleRecord.groupId === 1));
+  });
+});
+
+test('GranuleSearch with includeActiveStatus set to false does not return GranuleGroup info', async (t) => {
+  const { knex } = t.context;
+  const queryStringParameters = {
+    includeActiveState: 'false',
+  };
+  const dbSearch = new GranuleSearch({ queryStringParameters });
+  const response = await dbSearch.query(knex);
+  response.results.forEach((granuleRecord) => {
+    t.false('state' in granuleRecord);
+    t.false('groupId' in granuleRecord);
+  });
+});
+
+test('GranuleSearch supports GranuleGroup term search', async (t) => {
+  const { knex } = t.context;
+  let queryStringParameters = {
+    state: 'H',
+  };
+  let dbSearch = new GranuleSearch({ queryStringParameters });
+  let response = await dbSearch.query(knex);
+  response.results.forEach((granuleRecord) => {
+    t.true('state' in granuleRecord && granuleRecord.state === 'H');
+  });
+
+  queryStringParameters = {
+    groupId: '2',
+  };
+  dbSearch = new GranuleSearch({ queryStringParameters });
+  response = await dbSearch.query(knex);
+  response.results.forEach((granuleRecord) => {
+    t.true('groupId' in granuleRecord && granuleRecord.groupId === 2);
   });
 });